requirements.txt
openai 
python-pptx
python-dotenv
streamlit
googletrans==3.1.0a0